+++ plone/app/widgets/tests/test_at.py
@@ -62,6 +62,8 @@ class BaseWidgetTests(unittest.TestCase):
 
 class DateWidgetTests(unittest.TestCase):
 
+    layer = PLONEAPPWIDGETS_INTEGRATION_TESTING
+
     def setUp(self):
         from plone.app.widgets.at import DateWidget
         self.request = TestRequest(environ={'HTTP_ACCEPT_LANGUAGE': 'en'})
@@ -120,6 +122,8 @@ class DateWidgetTests(unittest.TestCase):
 
 class DatetimeWidgetTests(unittest.TestCase):
 
+    layer = PLONEAPPWIDGETS_INTEGRATION_TESTING
+
     def setUp(self):
         from plone.app.widgets.at import DatetimeWidget
         self.request = TestRequest(environ={'HTTP_ACCEPT_LANGUAGE': 'en'})
@@ -132,7 +136,7 @@ class DatetimeWidgetTests(unittest.TestCase):
         self.widget = DatetimeWidget()
         self.widget.pattern_options = {
             'date': {'firstDay': 0},
-            'time': {'interval': 15}
+            'time': {'interval': 5}
         }
 
     def test_widget(self):
@@ -166,10 +170,10 @@ class DatetimeWidgetTests(unittest.TestCase):
                                         u'Nov', u'Dec']
                     },
                     'time': {
+                        'interval': 5,
                         'placeholder': u'Enter time...',
                         'today': u'Today',
-                        'format': 'h:i a',
-                        'interval': 15
+                        'format': 'h:i a'
                     }
                 }
             },